Update simsum.R
Thank you for your great package. 
When applying it and comparing it to the results of my own code, I got different power/rejection-rates and checked your code. Your check of the control list seems to have a small error due to changes in names. Your checking if control$df is a number instead of control$power_df. This means that you cannot specify power_df at the moment. It's either set to the default NULL when you specify power_df in the control list or if you specify df it's returning an error for the name-control of the list.